+++ src/DateTime/Time.cpp
@@ -33,14 +33,14 @@
 #define NUM_NANOSECONDS_PER_DAY 86399999999999LL
 #define NANOSECONDS_PER_SECOND 1000000000LL
 
-static cass_int64_t php_driver_time_now_ns() {
+static cass_int64_t php_scylladb_time_now_ns() {
   cass_int64_t seconds;
   cass_int64_t nanoseconds;
 #if defined(__APPLE__) && defined(__MACH__)
   struct timeval ts {};
-  gettimeofday(&tv, NULL);
-  seconds = (cass_int64_t)tv.tv_sec;
-  nanoseconds = (cass_int64_t)tv.tv_usec * 1000;
+  gettimeofday(&ts, NULL);
+  seconds = (cass_int64_t)ts.tv_sec;
+  nanoseconds = (cass_int64_t)ts.tv_usec * 1000;
 #else
   struct timespec ts {};
   clock_gettime(CLOCK_REALTIME, &ts);
@@ -82,7 +82,7 @@ PHP_SCYLLADB_API zend_result php_scylladb_time_initialize(php_scylladb_time *sel
                                                           zend_string *nanosecondsStr,
                                                           zend_long nanoseconds) {
   if (nanosecondsStr == nullptr && nanoseconds == -1) {
-    self->time = php_driver_time_now_ns();
+    self->time = php_scylladb_time_now_ns();
     return SUCCESS;
   }