+++ CHANGELOG.md
@@ -1,10 +1,19 @@
-# Development
+
+# 1.3.10
+
+* fix:  math.cpp and math.h from utils added to CMakeLists.txt
+* feat: MacOS support for scripts/compile-php.sh
+* feat: Ext-ScyllaDB/CassandraDB support MacOS
+
+# 1.3.9
 
 * CMake used instead of autoconf and makefiles
 * Removed support for PHP 5.x
-* Version of the extension gets updated from CMakeLists.txt\
+* Version of the extension gets updated from CMakeLists.txt
 * Setup Pest Testing
-
+* Support for PHP 8.3
+* Default Dynamic Linking for LibUV and LibScyllaDB
+* Support for Cassandra 4.0 using LibCassandra from DataStax
 
 # 1.3.8